pyproject.toml
[project]
name = "nhmesh-telemetry"
version = "0.1.0"
description = ""
authors = [
    {name = "grleblanc",email = "greg@glsec.us"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "meshtastic (>=2.6.2,<3.0.0)",
    "protobuf-to-dict (>=0.1.0,<0.2.0)",
    "ruff (>=0.11.7,<0.12.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88

target-version = "py313"
indent-width = 2